# openapi.yaml
openapi: 3.0.0
info:
  title: SmileFile
  version: 1.0.0
servers:
  - url: /
paths:
  /patients:
    get:
      summary: Получить список пациентов
      parameters:
        - name: offset
          in: query
          description: Начальная точка отсчета для списка пациентов
          required: false
          schema:
            type: integer
            default: 0
        - name: limit
          in: query
          description: Ограничение на количество возвращаемых пациентов
          required: false
          schema:
            type: integer
            default: 1000
      responses:
        "200":
          description: Список пациентов
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Patient'
    post:
      summary: Добавить нового пациента
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewPatient'
      responses:
        "201":
          description: Пациент создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Patient'
        "400":
          description: Ошибка запроса
        "500":
          description: Внутренняя ошибка сервера
  /patients/{patientId}:
    get:
      summary: Получить конкретного пациента
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Информация о пациенте
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Patient'
        "404":
          description: Пациент не найден
        "500":
          description: Внутренняя ошибка сервера
    put:
      summary: Обновить информацию о пациенте
      description: |
        Список visits обновляется отдельно. Передавать список visits не нужно.
      parameters:
        - name: patientId
          description: Пользователя находим именно по этому id. Id в body не используется.
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Patient'
      responses:
        "200":
          description: Информация о пациенте обновлена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Patient'
        "400":
          description: Ошибка запроса
        "404":
          description: Пациент не найден
        "500":
          description: Внутренняя ошибка сервера
  /visits:
    get:
      summary: Получить список посещений
      description: |
        Посещения автоматически отсортированы по дате в порядке возрастания.
      parameters:
        - name: offset
          in: query
          description: Начальная точка отсчета для списка посещений
          required: false
          schema:
            type: integer
            default: 0
        - name: limit
          in: query
          description: Ограничение на количество возвращаемых посещений
          required: false
          schema:
            type: integer
            default: 1000
        - name: date
          in: query
          description: Дата для фильтрации посещений(если пустое, фильтра по дате нет)
          required: false
          schema:
            type: string
            format: date
            example: '2000-01-23'
      responses:
        "200":
          description: Список посещений
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Visit'

        "500":
          description: Внутренняя ошибка сервера
    post:
      summary: Создать новое посещение
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewVisit'
      responses:
        "201":
          description: Посещение создано
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Visit'
        "400":
          description: Ошибка запроса
        "500":
          description: Внутренняя ошибка сервера
  /visits/{visitId}:
    get:
      summary: Получить информацию о конкретном посещении
      parameters:
        - name: visitId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Информация о посещении
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Visit'
        "404":
          description: Посещение не найдено
        "500":
          description: Внутренняя ошибка сервера
    put:
      summary: Обновить информацию о посещении
      parameters:
        - name: visitId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Visit'
      responses:
        "200":
          description: Информация о посещении обновлена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Visit'
        "400":
          description: Ошибка запроса
        "404":
          description: Посещение не найдено
        "500":
          description: Внутренняя ошибка сервера
components:
  schemas:
    Patient:
      type: object
      properties:
        id:
          type: integer
          title: id
        name:
          type: string
          title: name
        contactDetails:
          type: string
          description: Контактные данные
          title: contactDetails
        lastAppointment:
          type: string
          format: date-time
          description: Дата последнего посещения
          title: lastAppointment
        visits:
          type: array
          items:
            $ref: '#/components/schemas/Visit'
          title: visits
      example:
        id: 1
        name: name
        contactDetails: contactDetails
        lastAppointment: "2000-01-23T04:56:07.000+00:00"
        visits:
          - id: 1
            patientId: 1
            date: "2000-01-23T04:56:07.000+00:00"
            treatment: treatment
            diagnosis: diagnosis
          - id: 2
            patientId: 1
            date: "2000-01-23T04:56:07.000+00:00"
            treatment: treatment
            diagnosis: diagnosis
    Visit:
      type: object
      properties:
        id:
          type: integer
          title: id
        patientId:
          type: integer
          title: patientId
        date:
          type: string
          format: date-time
          description: Дата посещения
          title: date
        diagnosis:
          type: string
          description: Диагноз
          title: diagnosis
        treatment:
          type: string
          description: Лечение
          title: treatment
      example:
        id: 1
        patientId: 1
        date: "2000-01-23T04:56:07.000+00:00"
        treatment: treatment
        diagnosis: diagnosis
    NewPatient:
      type: object
      properties:
        name:
          type: string
          title: name
        contactDetails:
          type: string
          description: Контактные данные
          title: contactDetails
        lastAppointment:
          type: string
          format: date-time
          description: Дата последнего посещения
          title: lastAppointment
      example:
        name: name
        contactDetails: contactDetails
        lastAppointment: "2000-01-23T04:56:07.000+00:00"
    NewVisit:
      type: object
      properties:
        patientId:
          type: integer
          title: patientId
        date:
          type: string
          format: date-time
          description: Дата посещения
          title: date
        diagnosis:
          type: string
          description: Диагноз
          title: diagnosis
        treatment:
          type: string
          description: Лечение
          title: treatment
      example:
        patientId: 1
        date: "2000-01-23T04:56:07.000+00:00"
        treatment: treatment
        diagnosis: diagnosis